% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getOutputFile.R
\name{getOutputFile}
\alias{getOutputFile}
\title{Get the output file produced by GMRFilm}
\usage{
getOutputFile()
}
\value{
outFile
}
\description{
Get the output file produced by GMRFilm
}
\examples{
### not run
# outFile = getOutputFile()
}
\keyword{keywords}

