% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeRunIt.R
\name{makeRunIt}
\alias{makeRunIt}
\title{Make a run-it cmd file for GMRFilm}
\usage{
makeRunIt(wrkDir = "C:/Temp/")
}
\arguments{
\item{wrkDir}{Working directory for GMRFilm simulation \code{wrkDir}}
}
\value{
none
}
\description{
Compute the NiKa and CuKa K-ratios using GMRFilm for a vector
of accelerating voltages for specified Ni and Cu thickness values.
Note: this corrects for continuum fluorescence
}
\examples{
### Not run
# makeRunIt()
}
\keyword{keywords}

