% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preClean.R
\name{preClean}
\alias{preClean}
\title{Pre-clean for GMRFilm}
\usage{
preClean(wrkDir = "C:/Temp/")
}
\arguments{
\item{wrkDir}{Working directory for GMRFilm simulation \code{wrkDir}}
}
\value{
none
}
\description{
Pre-clean for GMRFilm
}
\keyword{keywords}

