% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeInputCuOxCu.R
\name{writeInputCuOxCu.R}
\alias{writeInputCuOxCu.R}
\title{Write an input file for a GMRFilm simulation of CuOx on Cu}
\usage{
writeInputCuOxCu.R(tOx, wfOx, wfCu, rhoOx, e0, fPath, toa = 35)
}
\arguments{
\item{tOx}{Thickness of CuOx in nm \code{tOx}}

\item{wfOx}{weight fraction O in oxide layer \code{wfOx}}

\item{wfCu}{weight fraction Cu in oxide layer \code{wfCu}}

\item{rhoOx}{density of oxide layer \code{rhoOx}}

\item{e0}{A kV to simulate \code{e0}}

\item{fPath}{full path to the output file \code{fPath}}

\item{toa}{The detector take-off angle \code{toa}}
}
\value{
none
}
\description{
Write an input file for GMRFilm to construct an input file for
GMRFilm to compute the K-ratios for O Ka and Cu La for a
specified thickness of CuOx and Cu for a single e0 values.
}
\examples{
### not run
}
\keyword{keywords}

