% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moveIt.R
\name{moveIt}
\alias{moveIt}
\title{Move the output file for GMRFilm}
\usage{
moveIt(wrkDir = "C:/Temp/")
}
\arguments{
\item{wrkDir}{Working directory for GMRFilm simulation \code{wrkDir}}
}
\value{
none
}
\description{
Move the output file for GMRFilm
}
\examples{
### Not run
}
\keyword{keywords}

